+++ src/main/java/pl/asie/computronics/Computronics.java
@@ -82,11 +82,13 @@ import java.util.concurrent.Executors;
 
 @Mod(modid = Mods.Computronics, name = Mods.Computronics_NAME, version = "@VERSION@",
 	dependencies = "required-after:asielib@[0.4.0,);required-after:Forge@[10.13.2.1236,);"
-		+ "after:ComputerCraft;after:OpenComputers@[1.5.9,);after:OpenComputersAPI|Internal@[5.0.0,);"
-		+ "after:OpenComputers|Core;after:MineFactoryReloaded;after:RedLogic@[59.1.9,);after:ProjRed|Core;"
-		+ "after:nedocomputers;after:BuildCraft|Core@[6.4.1,);after:Railcraft@[9.5.0.0,);"
-		+ "after:gregtech@[MC1710];after:EnderIO@[1.7.10_2.2.7,);before:OpenPeripheralCore@[1.1,);"
-		+ "before:OpenPeripheralApi@[3.2,);after:Forestry@[3.5.3,);after:Waila@[1.5.10,);after:MekanismAPI|energy@[8.0.0,)")
+		+ "after:ComputerCraft;after:OpenComputers@[1.5.9,);after:nedocomputers;"
+		+ "before:OpenPeripheralCore@[1.1,);before:OpenPeripheralApi@[3.2,);"
+		+ "after:MineFactoryReloaded;after:RedLogic@[59.1.9,);after:ProjRed|Core;"
+		+ "after:BuildCraft|Core@[6.4.1,);after:Railcraft@[9.5.0.0,);"
+		+ "after:gregtech@[MC1710];after:EnderIO@[1.7.10_2.2.7,);"
+		+ "after:Forestry@[3.5.3,);after:Waila@[1.5.10,);"
+		+ "after:MekanismAPI|energy@[8.0.0,)")
 public class Computronics {
 	public Config config;
 	public Compat compat;