+++ src/main/java/pl/asie/computronics/reference/Config.java
@@ -2,8 +2,7 @@ package pl.asie.computronics.reference;
 
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.common.config.Configuration;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import pl.asie.computronics.Computronics;
 import pl.asie.lib.util.EnergyConverter;
 
 /**
@@ -38,7 +37,9 @@ public class Config {
 	public static boolean FORESTRY_BEES = true;
 	public static boolean BUILDCRAFT_STATION = true;
 
-	public static boolean OC_ROBOT_UPGRADES;
+	public static boolean OC_UPGRADE_CAMERA;
+	public static boolean OC_UPGRADE_CHATBOX;
+	public static boolean OC_UPGRADE_RADAR;
 	public static boolean OC_CARD_FX;
 	public static boolean OC_CARD_SPOOF;
 	public static boolean OC_CARD_SOUND;
@@ -89,7 +90,9 @@ public class Config {
 			CIPHER_WORK_CONSUMPTION = convertRFtoOC(
 				config.getFloat("cipherWorkConsumption", "power", 160.0f, 0.0f, 100000.0f, "How much base energy the Advanced Cipher Block should consume per encryption/decryption task. It will consume this value + 2*(number of characters in message)"));
 
-			OC_ROBOT_UPGRADES = config.get("enable.opencomputers", "robotUpgrades", true).getBoolean(true);
+			OC_UPGRADE_CAMERA = config.get("enable.opencomputers", "cameraUpgrade", true).getBoolean(true);
+			OC_UPGRADE_CHATBOX = config.get("enable.opencomputers", "chatboxUpgrade", true).getBoolean(true);
+			OC_UPGRADE_RADAR = config.get("enable.opencomputers", "radarUpgrade", true).getBoolean(true);
 			OC_CARD_FX = config.get("enable.opencomputers", "particleCard", true).getBoolean(true);
 			OC_CARD_SPOOF = config.get("enable.opencomputers", "spoofingCard", true).getBoolean(true);
 			OC_CARD_SOUND = config.get("enable.opencomputers", "soundCard", true).getBoolean(true);
@@ -128,16 +131,8 @@ public class Config {
 			CC_ALWAYS_FIRST = config.getBoolean("alwaysFirstPeripheral", "computercraft.multiperipheral", true, "If this is true, the Computronics MultiPeripheral system will almost always be the one recognized by ComputerCraft");
 			config.setCategoryComment("computercraft.multiperipheral", "If all of these options are set to true, Computronics will fix almost every conflict with multiple mods adding peripherals to the same block");
 			if(CC_OPEN_MULTI_PERIPHERAL && CC_ALL_MULTI_PERIPHERALS && CC_ALWAYS_FIRST) {
-				Logger cpx = LogManager.getLogger(Mods.Computronics_NAME);
-				Logger cc = LogManager.getLogger(Mods.ComputerCraft);
-				cpx.info("Hey, ComputerCraft! Guess what!");
-				cc.info("What?");
-				cpx.info("I fixed your peripheral system!");
-				cc.info("You did WHAT?!");
-				cpx.info("Multiple mods registering peripherals for the same block won't be a problem anymore!");
-				cc.info("Are you serious?");
-				cpx.info("Yes I am. Now be quiet and let Minecraft continue to load.");
-				cc.info("...");
+				Computronics.log.info("Multiperipheral system for ComputerCraft engaged. Hooray!");
+				Computronics.log.info("Multiple mods registering peripherals for the same block now won't be a problem anymore.");
 			}
 		}